+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 3.5.1
 ------
 
+**ENHANCEMENTS**
+- Add support for US isolated regions: us-iso-* and us-isob-*.
+
 3.5.0
 ------